Generate distributable
The distributable file can be generated and added a setup file.

+++ curve_plotter/controls.py
@@ -15,7 +15,7 @@ class Controls(wx.App):
 
     def initialize(self):
         if (self.steps_ui_edit.GetValue() != ''):
-            self.steps_value = float(self.steps_ui_edit.GetValue())
+            self.steps_value = int(self.steps_ui_edit.GetValue())
         
         self.plot_2d_checked = self.plot_2d_ui_checkbox.IsChecked()
